Move ItemComment into ItemDisplay package.
