--- packages/terra-clinical-item-comment/README.md
@@ -1,25 +1,0 @@
-# Terra Clinical Item View
-
-
-[![NPM version](http://img.shields.io/npm/v/terra-clinical-item-comment.svg)](https://www.npmjs.org/package/terra-clinical-item-comment)
-[![Build Status](https://travis-ci.org/cerner/terra-clinical.svg?branch=master)](https://travis-ci.org/cerner/terra-clinical)
-
-The Item Comment component creates a display for a comment icon and text.
-
-- [Getting Started](#getting-started)
-- [Documentation](https://github.com/cerner/terra-clinical/tree/master/packages/terra-clinical-item-comment/docs)
-- [LICENSE](#license)
-
-## Getting Started
-
-- Install from [npmjs](https://www.npmjs.com): `npm install terra-clinical-item-comment`
-
-## LICENSE
-
-Copyright 2017 Cerner Innovation, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
-
-&nbsp;&nbsp;&nbsp;&nbsp;http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
