--- packages/terra-clinical-item-comment/docs/README.md
@@ -1,22 +1,0 @@
-# Terra Clinical Item Comment
-
-The Item Comment component creates a display for a comment icon and text.
-
-## Getting Started
-
-- Install with [npmjs](https://www.npmjs.com):
-  - `npm install terra-clinical-item-comment`
-  - `yarn add terra-clinical-item-comment`
-
-## React Usage
-
-The React component can be initialized like so:
-```
-import Comment from 'terra-clinical-item-comment';
-
-<Comment
-  text="display text"
-  textStyle="attention"
-  isTruncated
-/>
-```
