--- packages/terra-clinical-item-comment/src/_mixins.scss
@@ -1,8 +1,0 @@
-@mixin terra-clinical-text-body {
-  font-size: 1.1428571428571428rem;
-  line-height: 1.4285714285714286;
-
-  @media screen and (min-width: 760px) {
-    font-size: 1rem;
-  }
-}
