--- packages/terra-clinical-item-comment/tests/nightwatch.conf.js
@@ -1,9 +1,0 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-
-const testSettings = require('../node_modules/terra-toolkit').testSettings;
-const resolve = require('path').resolve;
-const nightwatchConfiguration = require('../node_modules/terra-toolkit/lib/nightwatch.json');
-
-module.exports = (settings => (
-  testSettings(resolve('../../webpack.config'), settings)
-))(nightwatchConfiguration);
