+++ packages/terra-clinical-item-display/lib/ItemComment.scss
