+++ packages/terra-clinical-item-display/lib/ItemDisplay.js
@@ -18,6 +18,10 @@ require('terra-base/lib/baseStyles');
 
 require('./ItemDisplay.scss');
 
+var _ItemComment = require('./ItemComment');
+
+var _ItemComment2 = _interopRequireDefault(_ItemComment);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
@@ -86,5 +90,6 @@ var ItemDisplay = function ItemDisplay(_ref) {
 
 ItemDisplay.propTypes = propTypes;
 ItemDisplay.defaultProps = defaultProps;
+ItemDisplay.Comment = _ItemComment2.default;
 
 exports.default = ItemDisplay;
