+++ packages/terra-clinical-item-display/package.json
@@ -2,7 +2,7 @@
   "name": "terra-clinical-item-display",
   "main": "lib/ItemDisplay.js",
   "version": "0.0.0",
-  "description": "The Item Display component creates an display for text and an optional graphic.",
+  "description": "The Item Display component creates an display for text and an optional graphic. The Comment subcomponent creates a display for text with a comment icon.",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/cerner/terra-clinical.git"
@@ -33,6 +33,7 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "terra-base": "^0.x",
+    "terra-icon": "^0.x",
     "terra-mixins": "^1.0.0"
   },
   "scripts": {
