+++ packages/terra-clinical-item-display/src/ItemComment.jsx
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react';
 import classNames from 'classnames';
 import 'terra-base/lib/baseStyles';
 import CommentIndicator from 'terra-icon/lib/icon/IconComment';
-import ItemDisplay from 'terra-clinical-item-display';
+import ItemDisplay from './ItemDisplay';
 import './ItemComment.scss';
 
 const propTypes = {
