+++ packages/terra-clinical-item-display/src/ItemComment.scss
