+++ packages/terra-clinical-item-display/tests/nightwatch/display/IconItemDisplay.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-import ItemDisplay from '../../lib/ItemDisplay';
+import ItemDisplay from '../../../lib/ItemDisplay';
 
 export default () => <ItemDisplay icon={<img alt="graphic" />} />;
