+++ packages/terra-clinical-item-display/tests/nightwatch/display/ItemDisplayTests.jsx
