+++ packages/terra-clinical-item-view/tests/nightwatch/CommentItemView.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import ItemComment from 'terra-clinical-item-comment';
+import ItemDisplay from 'terra-clinical-item-display';
 import ItemView from '../../lib/ItemView';
 
 
-const comment1 = <ItemComment text="comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1" />;
+const comment1 = <ItemDisplay.Comment text="comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1" />;
 
 export default () => <ItemView comment={comment1} />;
