+++ packages/terra-clinical-site/package.json
@@ -30,7 +30,6 @@
     "terra-button": "^0.x",
     "terra-clinical-header": "^0.x",
     "terra-clinical-detail-view": "^0.x",
-    "terra-clinical-item-comment": "^0.x",
     "terra-clinical-item-display": "^0.x",
     "terra-clinical-item-view": "^0.x",
     "terra-clinical-error-view": "^0.x",
