+++ packages/terra-clinical-site/src/App.jsx
@@ -20,7 +20,6 @@ const App = props => (
       <li><Link to="/site/detail-view">Detail View</Link></li>
       <li><Link to="/site/error-view">Error View</Link></li>
       <li><Link to="/site/header">Header</Link></li>
-      <li><Link to="/site/item-comment">Item Comment</Link></li>
       <li><Link to="/site/item-display">Item Display</Link></li>
       <li><Link to="/site/item-view">Item View</Link></li>
       <li><Link to="/site/label-value-view">Label Value View</Link></li>
