+++ packages/terra-clinical-site/src/Index.jsx
@@ -7,7 +7,6 @@ import Home from './Home';
 // Examples
 import DetailViewExamples from './examples/detail-view/Index';
 import ErrorViewExamples from './examples/error-view/Index';
-import ItemCommentExamples from './examples/item-comment/Index';
 import ItemDisplayExamples from './examples/item-display/Index';
 import ItemViewExamples from './examples/item-view/Index';
 import HeaderExamples from './examples/header/Index';
@@ -20,9 +19,9 @@ import DetailViewTestRoutes from 'terra-clinical-detail-view/tests/nightwatch/De
 import DetailListTestRoutes from 'terra-clinical-detail-view/tests/nightwatch/DetailList/DetailListTestRoutes';
 import DetailListItemTestRoutes from 'terra-clinical-detail-view/tests/nightwatch/DetailListItem/ListItemTestRoutes';
 import ErrorViewTestRoutes from 'terra-clinical-error-view/tests/nightwatch/ErrorViewTestRoutes';
-import ItemCommentTestRoutes from 'terra-clinical-item-comment/tests/nightwatch/ItemCommentTestRoutes';
 import ItemViewTestRoutes from 'terra-clinical-item-view/tests/nightwatch/ItemViewTestRoutes';
-import ItemDisplayTestRoutes from 'terra-clinical-item-display/tests/nightwatch/ItemDisplayTestRoutes';
+import ItemDisplayTestRoutes from 'terra-clinical-item-display/tests/nightwatch/comment/ItemCommentTestRoutes';
+import ItemCommentTestRoutes from 'terra-clinical-item-display/tests/nightwatch/display/ItemDisplayTestRoutes';
 import HeaderTestRoutes from 'terra-clinical-header/tests/nightwatch/HeaderTestRoutes';
 import LabelValueViewTestRoutes from 'terra-clinical-label-value-view/tests/nightwatch/LabelValueViewTestRoutes';
 import NoDataViewTestRoutes from 'terra-clinical-no-data-view/tests/nightwatch/NoDataViewTestRoutes';
@@ -37,7 +36,6 @@ ReactDOM.render((
       <Route path="detail-view" component={DetailViewExamples} />
       <Route path="error-view" component={ErrorViewExamples} />
       <Route path="header" component={HeaderExamples} />
-      <Route path="item-comment" component={ItemCommentExamples} />
       <Route path="item-display" component={ItemDisplayExamples} />
       <Route path="item-view" component={ItemViewExamples} />
       <Route path="label-value-view" component={LabelValueViewExamples} />
