+++ packages/terra-clinical-site/src/TestLinks.jsx
@@ -9,8 +9,8 @@ const TestLinks = () => (
       <li><Link to="/tests/detail-list-tests">DetailList Tests</Link></li>
       <li><Link to="/tests/detail-list-item-tests">DetailListItem Tests</Link></li>
       <li><Link to="/tests/error-view-tests">Error View Tests</Link></li>
-      <li><Link to="/tests/item-comment-tests">ItemComment Tests</Link></li>
       <li><Link to="/tests/item-display-tests">ItemDisplay Tests</Link></li>
+      <li><Link to="/tests/item-comment-tests">ItemDisplay-Comment Tests</Link></li>
       <li><Link to="/tests/item-view-tests">ItemView Tests</Link></li>
       <li><Link to="/tests/header-tests">Header Tests</Link></li>
       <li><Link to="/tests/label-value-view-tests">Label Value View Tests</Link></li>
