--- packages/terra-clinical-site/src/examples/item-comment/DefaultItemComment.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import ItemComment from 'terra-clinical-item-comment';
-
-const DefaultItemComment = () => (<ItemComment text="Comment" />);
-
-export default DefaultItemComment;
