+++ packages/terra-clinical-site/src/examples/item-display/DefaultComment.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import ItemDisplay from 'terra-clinical-item-display';
+
+const DefaultComment = () => (<ItemDisplay.Comment text="Comment" />);
+
+export default DefaultComment;
