+++ packages/terra-clinical-site/src/examples/item-view/ItemViewAllTopAligned.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import ItemView from 'terra-clinical-item-view';
 import ItemDisplay from 'terra-clinical-item-display';
-import ItemComment from 'terra-clinical-item-comment';
 import IconPanelLeft from 'terra-icon/lib/icon/IconPanelLeft';
 import IconPanelRight from 'terra-icon/lib/icon/IconPanelRight';
 
@@ -17,7 +16,7 @@ const displays = [display1, display2, display3, display4, display5];
 const accessoryStart = <IconPanelLeft />;
 const accessoryEnd = <IconPanelRight />;
 
-const comment = <ItemComment text="test comment" />;
+const comment = <ItemDisplay.Comment text="test comment" />;
 
 export default () => <ItemView
   displays={displays}
